+++ rock/cli/command/job.py
@@ -18,8 +18,8 @@ class JobCommand(Command):
     async def _job_run(self, args: argparse.Namespace):
         # Import lazily to avoid pulling in bench/Harbor modules for bash-only uses
         from rock.sdk.bench.models.trial.config import RockEnvironmentConfig
+        from rock.sdk.job import Job
         from rock.sdk.job.config import BashJobConfig
-        from rock.sdk.job.facade import Job
 
         job_type = args.type or "bash"
